requirements.txt
opencv-python
pytesseract
numpy
pyspellchecker